requirements.txt
streamlit==1.32.0
openrouteservice==2.3.1
folium==0.15.1
streamlit-folium==0.15.1
geopy==2.4.1
python-dotenv==1.0.1 